Fix error stop test for flang
F23 standard for `execute_command_line` says optional argument
`cmdstat` is set to:

"a processor-dependent positive value if an error condition occurs,"

and flang's `execute_command_line` apparently sets cmdstat to a non-zero
value (3) when the command line returns a non-zero exit code.
This was breaking all the error stop tests under flang.

Change our error stop test to stop assuming a particular result in
`cmdstat` after executing a command that returns a non-zero exit code,
because that is not guaranteed by the standard.